+++ src/radical/pilot/controller/__init__.py
@@ -11,3 +11,4 @@ from pilot_manager_controller import PilotManagerController
 
 from input_file_transfer_worker import InputFileTransferWorker
 from output_file_transfer_worker import OutputFileTransferWorker
+